Kazaz/separate deploy scripts (#164)
* Seperate deploy of different chains to different folders

* Added script

* Added comment

* Update deploy-scripts/deploy-scripts-constants.ts
