+++ deploy-scripts/interfaces/deploy-scripts.interface.ts
@@ -1,0 +1,7 @@
+export interface TokenDeploymentData {
+  [key: string]: {
+    symbol: string;
+    name: string;
+    contract: string;
+  };
+}
