+++ deploy-scripts/rsk/3_deploy_uniswap_helpers.ts
@@ -1,7 +1,7 @@
 import type { HardhatRuntimeEnvironment } from "hardhat/types";
 import type { DeployFunction } from "hardhat-deploy/types";
-import * as constants from "../utils/Constants";
-import { formatAddress } from "../utils/DeployedContractsHelpers";
+import * as constants from "../../utils/Constants";
+import { formatAddress } from "../../utils/DeployedContractsHelpers";
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   console.log("starting deployment: uniswap helpers");
