+++ package.json
@@ -8,9 +8,8 @@
     "node": "hardhat node --no-deploy",
     "deploy:fork": "hardhat --network mainnetFork deploy",
     "deploy:fork:reset": "npm run deploy:fork -- --reset",
-    "deploy:rinkeby": "hardhat --network rinkeby deploy",
-    "deploy:kovan": "hardhat --network kovan deploy",
     "deploy:rsk": "hardhat --network rsk deploy",
+    "deploy:arbitrum-testnet": "hardhat --network arbitrumTestnet deploy",
     "initialize": "hardhat --network mainnetFork initialize:local",
     "update:oracles:fork": "hardhat --network mainnetFork update:oracles",
     "update:oracles:rsk": "hardhat --network rsk update:oracles",
